refactor: moved RecentUpdates to components
